+++ CHANGELOG.rst
@@ -11,15 +11,20 @@ develop
   --status_to_stdout
 - New setting `status_prefix` contains the string to emit prior to all
   status messages
+- New class `TomcatApplication`
 
 **Improved**
 - If we get an http redirect during `TomcatManager.connect()`, save the new
   url so we don't have to re-traverse the redirect on every command.
+- Interactive list command now can filter by application state, and has two
+  sort options.
 
 **Changed**
 
 - `TomcatManager._user` is now `TomcatManager.user`
 - `TomcatManager._url` is now `TomcatManager.url`
+- `TomcatManager.list()` now returns a list of `TomcatApplication` objects
+
 
 0.10.0 (2017-08-24)
 -------------------